+++ src/classes/Command/App/Install.php
@@ -88,7 +88,7 @@ class Hymn_Command_App_Install extends Hymn_Command_Abstract implements Hymn_Com
 		}
 
 		$installer	= new Hymn_Module_Installer( $this->client, $library );
-		$modules	= $relation->getOrder();
+		$modules	= $relation->getModulesOrderedByDependency();
 		foreach( $modules as $module ){
 			try{
 				$this->client->getFramework()->checkModuleSupport( $module );