+++ src/classes/Module/Graph.php
@@ -37,11 +37,11 @@
  */
 class Hymn_Module_Graph
 {
-	public const STATUS_EMPTY		= 0;
-	public const STATUS_CHANGED		= 1;
-	public const STATUS_LINKED		= 2;
-	public const STATUS_PRODUCED	= 3;
-	public const STATUS_DRAWN		= 4;
+	public const int STATUS_EMPTY		= 0;
+	public const int STATUS_CHANGED		= 1;
+	public const int STATUS_LINKED		= 2;
+	public const int STATUS_PRODUCED	= 3;
+	public const int STATUS_DRAWN		= 4;
 
 	/** @var		Hymn_Client				$client */
 	public Hymn_Client $client;
@@ -51,6 +51,8 @@ class Hymn_Module_Graph
 
 	/** @var		array					$nodes */
 	public array $nodes						= [];
+
+	protected Hymn_Structure_Graph $graph;
 
 	/** @var		object{quiet: bool, verbose: bool}	$flags */
 	protected object $flags;
@@ -62,6 +64,7 @@ class Hymn_Module_Graph
 	{
 		$this->client	= $client;
 		$this->library	= $library;
+		$this->graph	= new Hymn_Structure_Graph( $library );
 		$this->flags	= (object) [
 			'quiet'		=> (bool) ($this->client->flags & Hymn_Client::FLAG_QUIET ),
 			'verbose'	=> (bool) ($this->client->flags & Hymn_Client::FLAG_VERBOSE ),
@@ -77,46 +80,12 @@ class Hymn_Module_Graph
 	 *	@return		void
 	 */
 	public function addModule( Hymn_Structure_Module $module, int $level = 0 ): void
-  {
-//		if( version_compare( $this->client->getFramework()->getVersion(), '0.8.8.2', '<' ) )		//  framework is earlier than 0.8.8.2
-//			$module	= $this->library->getAvailableModule( $module->id );							//  load module using library
-
-		if( array_key_exists( $module->id, $this->nodes ) ){										//  module has been added already
-			if( $this->nodes[$module->id]->level < $level )											//  this time the level is deeper
-				$this->nodes[$module->id]->level	= $level;										//  store deeper level
-			return;																					//  exit without adding relations again
-		}
-		$this->nodes[$module->id]	= (object) [													//  add module to node list by module ID
-			'module'	=> $module,																	//  … store module data object
-			'level'		=> $level,																	//  … store load level
-			'in'		=> [],																		//  … store ingoing module links
-			'out'		=> [],																		//  … store outgoing module links
-		];
+	{
+		$this->graph->addModule( $module, $level );
+		$errors		= $this->graph->getErrors( TRUE );
+		if( [] !== $errors )
+			$this->client->outError( current( $errors ), Hymn_Client::EXIT_ON_RUN );
 		$this->status	= self::STATUS_CHANGED;														//  set internal status to "changed"
-		foreach( $module->relations->needs as $neededModuleId => $relation ){						//  iterate all modules linked as "needed"
-			//	 @todo remove this block after framework v0.8.8.2 is established
-			if( is_string( $relation ) ){															//  relation came from a reduced module source index
-				$neededModuleId	= $relation;														//  relation only holds module ID
-				$relation		= (object) [														//  simulate relation object
-					'type'		=> str_contains( $relation, '/' ) ? 'package' : 'module',			//  detect packages and modules
-					'source'	=> NULL,
-				];
-			}
-			if( $relation->type !== 'module' )
-			 	continue;
-			if( $relation->source ){
-				if( !$this->library->isAvailableModuleInSource( $neededModuleId, $relation->source ) ){
-					$message	= 'Module %s needs module %s from source %s, which is missing.';
-					$this->client->outError( vsprintf( $message, [
-						$module->id,
-						$neededModuleId,
-						$relation->source,
-					] ), Hymn_Client::EXIT_ON_RUN );
-				}
-			}
-			$neededModule	= $this->library->getAvailableModule( $neededModuleId, $relation->source );		//  get module data object from module library
-			$this->addModule( $neededModule, $level + 1 );											//  add this needed module with increased load level
-		}
 	}
 
 	/**
@@ -126,7 +95,7 @@ class Hymn_Module_Graph
 	 *	@access		public
 	 *	@return		array
 	 */
-	public function getOrder(): array
+	public function getModulesOrderedByDependency(): array
 	{
 		if( $this->status < self::STATUS_CHANGED )
 			throw new RuntimeException( 'No modules loaded' );
@@ -136,28 +105,30 @@ class Hymn_Module_Graph
 		/*  calculate maximum relation depth  */
 		$list	= [];
 		$max	= pow( 10, 8 ) - 1;
-		foreach( $this->nodes as $id => $node ){
-//			if( $this->flags->verbose && !$this->flags->quiet )
-//				$this->client->out( 'Check for loop: '.$node->module->id.' @ '.$node->module->sourceId );
-			$loop	= $this->checkForLoop( $node );
-			if( $loop ){
-				$this->client->outError( 'Module relation Loop found in module '.$loop->module->id.' @ '.$loop->module->sourceId );
+		$this->client->outVeryVerbose( 'Checking for loop in module relation graph...' );
+		foreach( $this->graph->getNodes() as $id => $node ){
+			/** @var ?object{module: Hymn_Structure_Module, modules: array<Hymn_Structure_Module>} $loop */
+			$loop	= $this->graph->checkForLoop( $node );
+			if( NULL !== $loop ){
+				$this->client->outError( 'Module relation loop found in module '.$loop->module->id.' @ '.$loop->module->sourceId );
 				foreach( array_values( $loop->modules ) as $nr => $item ){
 					$bullet	= str_pad( (string) ++$nr, 3, ' ', STR_PAD_LEFT );
 					$this->client->out( ' '.$bullet.'. '.$item->id.' @ '.$item->sourceId );
 				}
 				$this->client->outError( 'Please resolve loop, first!', Hymn_Client::EXIT_ON_RUN );
 			}
-			$edges	= $this->countModuleEdgesToRoot( $node );
+			$edges	= $this->graph->countModuleEdgesToRoot( $node );
 			$rand	= str_pad( (string) rand( 0, $max ), 8, '0', STR_PAD_LEFT );
 			$list[(float) $edges.'.'.$rand]	= $id;
 		}
 		krsort( $list );																			//  sort module order list
 
+		$this->client->outVeryVerbose( $this->client->getMemoryUsage( 'after calculating module order by dependency' ) );
+
 		/*  collect modules by installation order  */
 		$modules	= [];																		//  prepare empty module list
 		foreach( $list as $id )														//  iterate module order list
-			$modules[$id]	= $this->nodes[$id]->module;											//  collect module by installation order
+			$modules[$id]	= $this->graph->getNode( $id )->module;											//  collect module by installation order
 		return $modules;																			//  return list of modules by installation order
 	}
 
@@ -169,7 +140,7 @@ class Hymn_Module_Graph
 		$nodeStyle	= 'fontsize=9 shape=box color=black style=filled color="#00007F" fillcolor="#CFCFFF"';
 		$nodes	= [];
 		$edges	= [];
-		foreach( $this->nodes as $id => $node ){
+		foreach( $this->graph->getNodes() as $id => $node ){
 			$label		= 'label="'.$node->module->title.'"';
 			$nodes[]	= $node->module->id.' ['.$label.' '.$nodeStyle.'];';
 			foreach( $node->out as $out )
@@ -221,65 +192,14 @@ class Hymn_Module_Graph
 		return NULL;
 	}
 
-	/**
-	 *	Check for loop in module relations.
-	 *	@access		protected
-	 *	@param		object		$node		Node data object containing module and in and out relations
-	 *	@param		integer		$level		Counter of recursion level, 0 by default.
-	 *	@return		object|NULL				Object if looping node or null if no loop found
-	 */
-	protected function checkForLoop( object $node, int $level = 0, array $steps = [] ): ?object
-  {
-		if( array_key_exists( $node->module->install->path, $steps ) )								//  been in this module in before
-			return (object) array(																	//  return loop data ...
-				'module'	=> $node->module,														//  ... containing looping module
-				'modules'	=> $steps																//  ... and the module chain
-			);
-		$steps[$node->module->install->path]	= $node->module;									//  note this module in module chain
-		if( count( $node->in ) ){																	//  there are relations
-			foreach( $node->in as $parent ){														//  iterate these relations
-				$parent	= $this->nodes[$parent->id];												//  shortcut parent module
-				$loop	= $this->checkForLoop( $parent, $level + 1, $steps );						//  recurse for related node
-				if( $loop )																			//  found loop
-					return $loop;																	//  return looping node to prior rounds
-			}
-		}
-		return NULL;																				//  no loop found
-	}
-
-	protected function countModuleEdgesToRoot( object $node, int $level = 0 ): int
-	{
-		$count	= $level;
-		if( count( $node->in ) ){
-			$ways	= [];
-			foreach( $node->in as $parent )
-				$ways[]	= $this->countModuleEdgesToRoot( $this->nodes[$parent->id], $level + 1 );
-			$count	= max( $ways );
-		}
-		return $count;
-	}
-
 	protected function realizeRelations(): void
 	{
-		/*  count ingoing and outgoing module links  */
-		foreach( $this->nodes as $id => $node ){													//  iterate all nodes
-			foreach( $node->module->relations->needs as $neededModuleId => $relation ){				//  iterate all needed modules of node
-				//	 @todo remove this block after framework v0.8.8.2 is established
-				if( is_string( $relation ) ){														//  relation came from a reduced module source index
-					$neededModuleId	= $relation;													//  relation only holds module ID
-					$relation		= (object) [													//  simulate relation object
-						'type'		=> str_contains( $relation, '/' ) ? 'package' : 'module',		//  detect packages and modules
-						'source'	=> NULL,
-					];
-				}
-				if( $relation->type === 'module' ){
-					$this->nodes[$id]->out[$neededModuleId]	= $this->nodes[$neededModuleId];		//  note outgoing link on this node
-					$this->nodes[$neededModuleId]->in[$id]	= $node->module;						//  note ingoing link on the needed node
-				}
-			}
-		}
+		if( self::STATUS_LINKED <= $this->status )
+			return;
+		$this->graph->realizeRelations();
 		$this->status	= self::STATUS_LINKED;
 		if( $this->flags->verbose && !$this->flags->quiet )
 			$this->client->outVeryVerbose( "Found ".count( $this->nodes )." modules." );
+		$this->client->outVeryVerbose( $this->client->getMemoryUsage( 'after realizing module relations' ) );
 	}
 }