+++ src/classes/Module/Library.php
@@ -62,9 +62,10 @@ class Hymn_Module_Library
 		$this->available	= new Hymn_Module_Library_Available( $client );
 	}
 
-	public function addSource( string $sourceId, string $path, string $type, bool $active = TRUE, string $title = NULL ): void
+	public function addSource( string $sourceId, string $path, string $type, bool $active = TRUE, string $title = NULL ): static
 	{
 		$this->available->addSource( $sourceId, $path, $type, $active, $title );
+		return $this;
 	}
 
 	public function getActiveSources( bool $withModules = FALSE ): array
@@ -114,6 +115,12 @@ class Hymn_Module_Library
 	public function getSource( string $sourceId, bool $withModules = FALSE ): Hymn_Structure_Source
 	{
 		return $this->available->getSource( $sourceId, $withModules );
+	}
+
+	public function addModuleToSource( Hymn_Structure_Module $module, Hymn_Structure_Source $source ): static
+	{
+		$this->available->addModuleToSource( $module, $source );
+		return $this;
 	}
 
 	/**