+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v1.0.2
 - Fix structure migration bug in module update.
 - Add log output of updatable modules.
+- Add graph structure for module relations.
 
 ## v1.0.1
 - Move tool config files to own folder.